changed variable and group names, refactored code from view controller into funcitons in userCellView class
