+++ UsersApp/Views/userCellView.xib
@@ -4,13 +4,12 @@
     <dependencies>
         <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="21679"/>
         <capability name="Safe area layout guides" minToolsVersion="9.0"/>
-        <capability name="System colors in document resources" minToolsVersion="11.0"/>
         <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
     </dependencies>
     <objects>
         <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
         <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-        <view contentMode="scaleToFill" id="iN0-l3-epB" customClass="userCell" customModule="UsersApp" customModuleProvider="target">
+        <view contentMode="scaleToFill" id="iN0-l3-epB" customClass="userCellView" customModule="UsersApp" customModuleProvider="target">
             <rect key="frame" x="0.0" y="0.0" width="393" height="80"/>
             <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
             <subviews>
@@ -47,7 +46,7 @@
                 </label>
             </subviews>
             <viewLayoutGuide key="safeArea" id="vUN-kp-3ea"/>
-            <color key="backgroundColor" systemColor="systemIndigoColor"/>
+            <color key="backgroundColor" white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
             <constraints>
                 <constraint firstAttribute="trailing" secondItem="e2k-dP-kSb" secondAttribute="trailing" constant="10" id="2Sg-Gk-XZd"/>
                 <constraint firstItem="t9d-qf-Tph" firstAttribute="top" secondItem="2MI-Eu-bAv" secondAttribute="bottom" constant="10" id="6l3-hw-h1a"/>
@@ -64,14 +63,9 @@
                 <outlet property="emailLabel" destination="t9d-qf-Tph" id="8v2-2p-H4k"/>
                 <outlet property="nameLabel" destination="2MI-Eu-bAv" id="OZr-1r-TvI"/>
                 <outlet property="timeLabel" destination="e2k-dP-kSb" id="By4-zQ-K4y"/>
-                <outlet property="userImage" destination="bTq-36-fXI" id="x5A-tz-cuv"/>
+                <outlet property="userImageView" destination="bTq-36-fXI" id="x5A-tz-cuv"/>
             </connections>
             <point key="canvasLocation" x="75.572519083969468" y="20.422535211267608"/>
         </view>
     </objects>
-    <resources>
-        <systemColor name="systemIndigoColor">
-            <color red="0.34509803921568627" green="0.33725490196078434" blue="0.83921568627450982" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
-        </systemColor>
-    </resources>
 </document>
